Fifth Commit
After reducing the file size of the political data csv, there were
still problems syncing the folder with the new project between the
local version of the repo and the remote version. So, as a way to test
a suspected remedy for that problem, I re-foldered the content. This
commit is for adding the same content (but contained in a new folder)
to the repo for trying to sync in this way instead.